+++ packages/plugin/src/Services/Pro/RecaptchaService.php
@@ -14,6 +14,8 @@ use Solspace\Freeform\Models\Settings;
 
 class RecaptchaService extends Component
 {
+    private $lastError;
+
     public function validateRecaptchaV2Checkbox(ValidateEvent $event)
     {
         if ($this->isRecaptchaTypeSkipped(Settings::RECAPTCHA_TYPE_V2_CHECKBOX)) {
@@ -39,7 +41,7 @@ class RecaptchaService extends Component
                 $message = $this->getSettings()->recaptchaErrorMessage;
                 $event->getForm()->addError(Freeform::t($message ?: 'Please verify that you are not a robot.'));
             } else {
-                $event->getForm()->markAsSpam(SpamReason::TYPE_RECAPTCHA, 'ReCAPTCHA invisible validation failed');
+                $event->getForm()->markAsSpam(SpamReason::TYPE_RECAPTCHA, 'ReCAPTCHA - '.$this->lastError);
             }
         }
     }
@@ -56,7 +58,7 @@ class RecaptchaService extends Component
                 $message = $this->getSettings()->recaptchaErrorMessage;
                 $event->getForm()->addError(Freeform::t($message ?: 'Your submission could not be processed.'));
             } else {
-                $event->getForm()->markAsSpam(SpamReason::TYPE_RECAPTCHA, 'ReCAPTCHA score check failed');
+                $event->getForm()->markAsSpam(SpamReason::TYPE_RECAPTCHA, 'ReCAPTCHA - '.$this->lastError);
             }
         }
     }
@@ -131,7 +133,39 @@ class RecaptchaService extends Component
             $minScore = min(1, $minScore);
             $minScore = max(0, $minScore);
 
+            if ($result['score'] < $minScore) {
+                $this->lastError = 'Score check failed with ['.$result['score'].']';
+            }
+
             return $result['score'] >= $minScore;
+        }
+
+        if (!$result['success']) {
+            $errorCodes = $result['error-codes'];
+
+            if (\in_array('missing-input-secret', $errorCodes, true)) {
+                $this->lastError = 'The secret parameter is missing.';
+            }
+
+            if (\in_array('invalid-input-secret', $errorCodes, true)) {
+                $this->lastError = 'The secret parameter is invalid or malformed.';
+            }
+
+            if (\in_array('missing-input-response', $errorCodes, true)) {
+                $this->lastError = 'The response parameter is missing.';
+            }
+
+            if (\in_array('invalid-input-response', $errorCodes, true)) {
+                $this->lastError = 'The response parameter is invalid or malformed.';
+            }
+
+            if (\in_array('bad-request', $errorCodes, true)) {
+                $this->lastError = 'The request is invalid or malformed.';
+            }
+
+            if (\in_array('timeout-or-duplicate', $errorCodes, true)) {
+                $this->lastError = 'The response is no longer valid: either is too old or has been used previously.';
+            }
         }
 
         return $result['success'];